fix: incorrect selection coordinates when input grows (#491)
## 📜 Description

Fixed incorrect selection coordinates when text gets moved to a new line
and text input grows.

## 💡 Motivation and Context

The problem happens because our text is bigger and can not fit into
current layout. If we read coordinates we get `+Inf` and of course it's
not reasonable to send such events to JS.

To fix this problem we need to wait for layout update (next frame) and
send event only after that. This is exactly what i did in this PR:

```swift
if textView.canSelectionFitIntoLayout {
  updateSelectionPosition(textInput: textView, sendEvent: onSelectionChange)
} else {
  // when multiline input grows we need to wait for layout to be updated
  // otherwise start/end positions will be incorrect (0/-1)
  DispatchQueue.main.asyncAfter(deadline: .now() + UIUtils.nextFrame) {
    updateSelectionPosition(textInput: textView, sendEvent: self.onSelectionChange)
  }
}
```

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/489

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### iOS

- added `canSelectionFitIntoLayout` property extension to `UITextInput`;
- use `canSelectionFitIntoLayout` to detect whether we can send event
immediately or need to wait for one frmae for layout to be updated and
send event only after that;

## 🤔 How Has This Been Tested?

Tested manually using code sample provided in the issue.

## 📸 Screenshots (if appropriate):

|Before|After|
|-------|-----|

|![image](https://github.com/kirillzyusko/react-native-keyboard-controller/assets/22820318/4a4cff0e-a478-491e-bdd9-7f2164658cf9)|![image](https://github.com/kirillzyusko/react-native-keyboard-controller/assets/22820318/4e4ed628-a811-42f6-bf6e-a6d349e50fde)|

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed