fix: handle `contextMenuHidden` prop (#498)
## 📜 Description

Find a real `TextInput` instead of its parent when `contextMenuHidden`
prop is specified.

## 💡 Motivation and Context

If `contextMenuHidden` is specified then `UIResponder.current` returns
the parent of the actual `TextInput`. In case of **paper** element will
be:
- `RCTSinglelineTextInputView` (single line)
- `RCTMultilineTextInputView` (multi line)

On **fabric** it'll always be:
- `RCTTextInputComponentView`

Casting view to one of the following classes can be expensive in terms
of supporting RN versions - RN team can change class names/files and it
will not be considered as a breaking change (because it's not public
API), but we'll need to make conditional imports/do conditional casting
etc.

But in fact we just need to check that current responder belongs to
these classes and if yes then take the first children (with casting to
`UIView`). That's what I did in this PR.

I think this PR shouldn't bring any breaking changes, since we just
handle one small corner case.

Also we had to re-work `KeyboardController.dismiss` functionality - we
need to use our implementation for detecting first responder, so I had
to add `@objc` annotation to use my extension in ObjC code.

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/496

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->


### iOS

- detect if first responder is parent container and return its children
as first responder;

## 🤔 How Has This Been Tested?

Tested on CI + updated example app to be sure we verify newly added
code.

## 📸 Screenshots (if appropriate):

<img
src="https://github.com/kirillzyusko/react-native-keyboard-controller/assets/22820318/e4ddb6f7-ceca-4608-b567-f0d818f5e457"
height="600">

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed