feat: react on selection changes in `KeyboardAwareScrollView` (#508)
## 📜 Description

`KeyboardAwareScrollView` is reacting on text selection now.

## 💡 Motivation and Context

Sometimes users may scroll up, and then they may want to select part of
the text. If the text is hidden under the keyboard, then they do not
need what to select, they have to scroll themself, and it makes UX
worse. In this PR we automatically detect selection event and scroll
down to make sure selected text is visible.

Initially I wanted to scroll in the same manner as we do for text change
- i.e. just scroll to make TextInput to be above the keyboard. While it
works I thought on how it could be improved: since we have selection
coordinates we can re-use them and rely on them instead of the height of
the input. In this case we will scroll only for selection distance. I
experimented a little bit with this approach, it works so I decided to
stick to it.

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/420

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### JS

- create common `scrollFromCurrentPosition` function;
- re-use `scrollFromCurrentPosition` in `onChangeText` and
`onSelectionChange` handlers;

### iOS

- update layout on selection change event;

### Android

- update layout on selection change event;

### Docs

- mention new lifecycle when we update layout;

### E2E

- cover new functionality by new test;

## 🤔 How Has This Been Tested?

Tested on iPhone 15 Pro, Pixel 2 (API 28, API 31).

## 📸 Screenshots (if appropriate):


![image](https://github.com/user-attachments/assets/b894a1c3-f6bc-474d-8b97-b621ab4a98f1)

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed