e2e: stable selection test on API 31 (#513)
## 📜 Description

In
https://github.com/kirillzyusko/react-native-keyboard-controller/pull/508
I added new test. But it turned out that this test is not stable.

## 💡 Motivation and Context

The problem comes from the asynchronous keyboard resize/focus changed on
Android. Because of that at certain point of time in test the input can
be significantly above the keyboard. If we sue static interval (60px)
then sometimes it may be not enough to scroll it down, since selection
in this case will not scroll anything (because the distance between
selection and keyboard can be already > 50px).

To fix this problem I added a helper
`scrollUpUntilElementIsBarelyVisible`. This method will scroll down and
will assure, that element is hidden by the keyboard and then will
slightly scroll up. With this method we can assure, that position after
selection will be stable and subsequent tests will not fail.

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### E2E

- added `scrollUpUntilElementIsBarelyVisible` and started to use it in
one test;

## 🤔 How Has This Been Tested?

Tested locally and on CI.

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed