fix: pick up proper `ViewController` for `KeyboardToolbar` when modal is shown (#514)
## 📜 Description

Fixes incorrect button state if `KeyboardToolbar` is used inside the
modal window.

## 💡 Motivation and Context

The problem was because of the fact that when modal is shown we still
retrieve the `rootView`. As a result we searched for other inputs (not
these that user currently see) and the state of button was broken.

To overcome this problem I decided to search for current active view
controller (i. e. modal if it's shown). For that I created an
`UIApplication` extension. It worked, but I noticed one bug (actually
e2e tests caught that) - when contact modal is shown and you select the
contact, then you move focus to the next input and both arrows should be
active. However the state gets broken. It happens because of the fact
that we try to search for inputs in modal that has been dismissed. To
fix this problem I decided to take only visible modals into
consideration and fixed the problem in
https://github.com/kirillzyusko/react-native-keyboard-controller/pull/514/commits/095c167b942297aa130455c9a84aa01e5f842c46

Closes
https://github.com/kirillzyusko/react-native-keyboard-controller/issues/510

## 📢 Changelog

### iOS

- added new `UIApplication` extension;
- linked new extension to unit tests project;
- started to use `UIApplication.topViewController` instead of
`rootViewController`.

## 🤔 How Has This Been Tested?

tested manually on iPhone 15 Pro + e2e tests.

## 📸 Screenshots (if appropriate):

<img
src="https://github.com/user-attachments/assets/327d0ce5-d70e-44e1-b332-74ee73ad06f3"
height="600">

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed