ci: allow manual publishing (#521)
## 📜 Description

Allow manual run for publishing to npm action.

## 💡 Motivation and Context

I wanted to have "verified" badge, but forgot to run proper permissions
to the job. As a result the job failed and I can not publish a package
with new permission
(https://github.com/kirillzyusko/react-native-keyboard-controller/pull/520),
because the job is heavily depends on the fact that **new** tag must be
published. In oreder not to create a new tag I decided to allow manual
run for the job (as a last resort) 🙃

## 📢 Changelog

<!-- High level overview of important changes -->
<!-- For example: fixed status bar manipulation; added new types
declarations; -->
<!-- If your changes don't affect one of platform/language below - then
remove this platform/language -->

### CI

- allow manual runs for `publish.yml` job;
- rename `build` -> `publish` job;

## 🤔 How Has This Been Tested?

Can be tested only after merge 🙈 

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed