feat: allow keyboard toolbar button callbacks to cancel default actions (#532)
## 📜 Description

This PR allows the `onNextCallback`, `onPrevCallback`, and
`onDoneCallback` functions to cancel the default actions of the
next/prev/done buttons on `KeyboardToolbar`. This is done by passing the
React Native press event instance (`GestureResponderEvent`) to the
callbacks, allowing them to cancel the event with
`event.preventDefault()`. If no callback has been specified, or if the
callback does not invoke `event.preventDefault()`, the default action is
executed, which maintains current behaviour. There are no breaking
changes included by this PR.

## 💡 Motivation and Context

Imagine a form with a text field, a select field, and another text field
(when clicking the select field, a modal appears, allowing the user to
pick a value). In this particular situation, if the first text field is
selected, and we click the arrow down button, we would want to open the
select field modal instead of jumping to the second text field.

This PR makes this possible cleanly by providing a way to cancel the
default action, for example:
```ts
function onPressNext(event: GestureResponderEvent) {
    if (isFirstInputFocused()) event.preventDefault();
}
``` 

## 📢 Changelog

### JS

- ~Created new `KeyboardToolbarButtonPressEvent` class to represent a
keyboard toolbar button press event. For now, only contains a
`cancelled: boolean` property and a `cancel()` method. In the future,
this can be extended to add more event info, such as what input is
presently focused, the React Native press event which triggered this,
etc.~
- Modified the `onPressNext`, `onPressPrev`, and `onPressDone` handlers
so that they first invoke the user-specified callback and pass along the
`GestureResponderEvent` event instance, and only then execute the
default behaviour, if the event hasn't been cancelled.

## 🤔 How Has This Been Tested?

~I have only tested this manually and locally for now. I could write
tests but not too familiar with the existing tooling.~

TODO: to be tested following some changes.

## 📸 Screenshots (if appropriate):

N/A

## 📝 Checklist

- [ ] CI successfully passed
- [ ] I added new mocks and corresponding unit-tests if library API was
changed

---------

Co-authored-by: kirillzyusko <zyusko.kirik@gmail.com>