refactor: reusable `TextInput` protocol (#696)
## 📜 Description

Put `TextInput` protocol in a separate directory to highlight that it's
common protocol that can be re-used by other project files (not only
traversal group).

## 💡 Motivation and Context

In preparation for `offset` functionality of interactive keyboard I
discovered that we need to have common `UITextField` and `UITextView`
interface to be able to manage some props without additional
`if`-statements.

We already have our own `TextInput` protocol, so I thought we could
benefit from it and extend it for future functionality. The only one
thing is that now `TextInput` is associated with `traversal` group, but
since it can be re-used by other groups I thought it would be better to
move the file into its own separate group. That's what I did in this PR
🙂

## 📢 Changelog

### iOS

- move `TextInput` file from `traversal` to `protocol` group;

## 🤔 How Has This Been Tested?

Tested on CI.

## 📝 Checklist

- [x] CI successfully passed
- [x] I added new mocks and corresponding unit-tests if library API was
changed